refactor(refinement): move objection collection to its own module for size guard
Adding collect_adversary_round_objections plus its tests pushed
refinement_outcome.rs past the 51200-byte size-guard limit. Move the pure
function and its five unit tests into a new refinement_objections.rs sibling
module (same split rationale as refinement_outcome vs refinement_dispatch);
run-scoping predicate entry_in_current_run stays in refinement_outcome.
