+++ server/crates/djinn-coordinator/src/lib.rs
@@ -135,6 +135,7 @@ pub(crate) mod refinement;
 pub(crate) mod refinement_dispatch;
 #[cfg(test)]
 mod refinement_e2e_evidence_regression_tests;
+mod refinement_objections;
 mod refinement_outcome;
 mod refinement_recovery;
 pub mod rules;
